changed types from unsigned long long to uint64;int32_t to int32 and uint32_t to uint32 (these types are defined in UE4)

+++ Source/rclUE/Private/ROS2ActionClient.cpp
@@ -115,9 +115,9 @@ void UROS2ActionClient::CancelActionRequest()
     UE_LOG(LogROS2Action, Log, TEXT("A. Action Client - Send cancel action request (%s)"), *__LOG_INFO__);
 	action_msgs__srv__CancelGoal_Request* cancel_request = (action_msgs__srv__CancelGoal_Request*) Action->GetCancelRequest();
 	float CancelTime = UGameplayStatics::GetTimeSeconds(GWorld);
-	cancel_request->goal_info.stamp.sec = (int32_t)CancelTime;
-	unsigned long long ns = (unsigned long long)(CancelTime * 1000000000.0f);
-	cancel_request->goal_info.stamp.nanosec = (uint32_t)(ns - (cancel_request->goal_info.stamp.sec * 1000000000ul));
+	cancel_request->goal_info.stamp.sec = (int32)CancelTime;
+	uint64 ns = (uint64)(CancelTime * 1000000000.0f);
+	cancel_request->goal_info.stamp.nanosec = (uint32)(ns - (cancel_request->goal_info.stamp.sec * 1000000000ul));
 
 	int64_t Seq;
 	RCSOFTCHECK(rcl_action_send_cancel_request(&client, Action->GetCancelRequest(), &Seq));
