+++ Source/rclUE/Private/ROS2ActionServer.cpp
@@ -84,9 +84,9 @@ void UROS2ActionServer::SendGoalResponse()
 	ue4_interfaces__action__Fibonacci_SendGoal_Response* GoalResponse = (ue4_interfaces__action__Fibonacci_SendGoal_Response*)Action->GetGoalResponse();
 	GoalResponse->accepted = true;
 	float TimeOfResponse = UGameplayStatics::GetTimeSeconds(GWorld);
-	GoalResponse->stamp.sec = (int32_t)TimeOfResponse;
-	unsigned long long ns = (unsigned long long)(TimeOfResponse * 1000000000.0f);
-	GoalResponse->stamp.nanosec = (uint32_t)(ns - (GoalResponse->stamp.sec * 1000000000ul));
+	GoalResponse->stamp.sec = (int32)TimeOfResponse;
+	uint64 ns = (uint64)(TimeOfResponse * 1000000000.0f);
+	GoalResponse->stamp.nanosec = (uint32)(ns - (GoalResponse->stamp.sec * 1000000000ul));
 
 	RCSOFTCHECK(rcl_action_send_goal_response(&server, &goal_req_id, Action->GetGoalResponse()));
 }
@@ -99,9 +99,9 @@ void UROS2ActionServer::ProcessAndSendCancelResponse()
 
 	rcl_action_cancel_request_t cancel_request = rcl_action_get_zero_initialized_cancel_request();
 	float TimeOfCancelProcess = UGameplayStatics::GetTimeSeconds(GWorld);
-	cancel_request.goal_info.stamp.sec = (int32_t)TimeOfCancelProcess;
-	unsigned long long ns = (unsigned long long)(TimeOfCancelProcess * 1000000000.0f);
-	cancel_request.goal_info.stamp.nanosec = (uint32_t)(ns - (cancel_request.goal_info.stamp.sec * 1000000000ul));
+	cancel_request.goal_info.stamp.sec = (int32)TimeOfCancelProcess;
+	uint64 ns = (uint64)(TimeOfCancelProcess * 1000000000.0f);
+	cancel_request.goal_info.stamp.nanosec = (uint32)(ns - (cancel_request.goal_info.stamp.sec * 1000000000ul));
 	rcl_action_cancel_response_t cancel_response = rcl_action_get_zero_initialized_cancel_response();
 	RCSOFTCHECK(rcl_action_process_cancel_request(&server, &cancel_request, &cancel_response));
